Move TelemetryTokenCacheAccessor into its own class file
